requirements.txt
google-genai
requests
urllib3